+++ src/Analysis/DoctrineConstantListProvider.php
@@ -43,11 +43,7 @@ final class DoctrineConstantListProvider implements ConstantListProviderInterfac
         $constants = [];
 
         try {
-            $items = $this->managerRegistry->getRepository(Structures\Constant::class)->createQueryBuilder('entity')
-                ->select('entity')
-                ->andWhere('entity.structure IS NULL')
-                ->getQuery()
-                ->execute();
+            $items = $this->managerRegistry->getRepository(Structures\Constant::class)->findAll();;
         } catch (DriverException $e) {
             throw new RuntimeException($e->getMessage(), 0, $e);
         }